Moved rest of util functions
