+++ client/pages/edit/history/[resource]/[id].tsx
@@ -1,7 +1,8 @@
 import React, { useEffect, useState } from 'react';
 import { GetServerSidePropsContext, InferGetServerSidePropsType } from 'next';
 import { useRouter } from 'next/router';
-import { calculateEditDate, getParams } from '../../../../src/util';
+import { getParams } from '../../../../src/util/miscUtilUtil';
+import { calculateEditDate } from '../../../../src/util/datetime';
 import { darkSwitch } from '../../../../src/util/cssUtil';
 import { getHistory } from '../../../../src/api';
 import type { History, Resource } from '../../../../src/types';
