+++ client/pages/edit/index.tsx
@@ -2,7 +2,8 @@ import React, { useState, useEffect } from 'react';
 import { useRouter } from 'next/router';
 import { GetServerSidePropsContext, InferGetServerSidePropsType } from 'next';
 import { History, Resource, HistoryData, AccessLevel } from '../../src/types';
-import { calculateEditDate, getTokenFromCookies, redirect } from '../../src/util';
+import { getTokenFromCookies, redirect } from '../../src/util/miscUtilUtil';
+import { calculateEditDate } from '../../src/util/datetime';
 import { darkSwitch } from '../../src/util/cssUtil';
 import { getHistoryList, getUserInfo } from '../../src/api';
