+++ client/pages/events/reservations/[[...date]].tsx
@@ -3,7 +3,9 @@ import { GetServerSidePropsContext, InferGetServerSidePropsType } from 'next';
 import dayjs, { Dayjs } from 'dayjs';
 import { getReservationList } from '../../../src/api';
 import { AccessLevel } from '../../../src/types';
-import { getAccessLevel, parseDateParams, parseReservations } from '../../../src/util';
+import { getAccessLevel } from '../../../src/util/miscUtilUtil';
+import { parseReservations } from '../../../src/util/dataParsing';
+import { parseDateParams } from '../../../src/util/datetime';
 
 import Box from '@mui/material/Box';
 import TextField from '@mui/material/TextField';
