+++ client/pages/index.tsx
@@ -2,7 +2,9 @@ import React, { useState, useEffect } from 'react';
 import { GetServerSidePropsContext, InferGetServerSidePropsType } from 'next';
 import dayjs from 'dayjs';
 import { getPublicEventList } from '../src/api';
-import { parsePublicEventList, isSameDate, getAccessLevel } from '../src/util';
+import { getAccessLevel } from '../src/util/miscUtilUtil';
+import { parsePublicEventList } from '../src/util/dataParsing';
+import { isSameDate } from '../src/util/datetime';
 import { darkSwitchGrey } from '../src/util/cssUtil';
 import type { SxProps, Theme } from '@mui/material';
