+++ client/pages/profile/index.tsx
@@ -3,7 +3,7 @@ import { useRouter } from 'next/router';
 import { GetServerSidePropsContext, InferGetServerSidePropsType } from 'next';
 import Cookies from 'universal-cookie';
 import { getBackendUrl, getUserInfo, postLogin } from '../../src/api';
-import { createPopupEvent } from '../../src/util';
+import { createPopupEvent } from '../../src/util/constructors';
 import { darkSwitchGrey } from '../../src/util/cssUtil';
 import type { PopupEvent } from '../../src/types';
