+++ client/pages/volunteering/[id].tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { GetServerSidePropsContext, InferGetServerSidePropsType } from 'next';
 import { useRouter } from 'next/router';
 import { getVolunteering } from '../../src/api';
-import { getAccessLevel, getParams } from '../../src/util';
+import { getAccessLevel, getParams } from '../../src/util/miscUtilUtil';
 import { darkSwitchGrey } from '../../src/util/cssUtil';
 
 import Container from '@mui/material/Container';
