+++ client/pages/volunteering/index.tsx
@@ -1,7 +1,7 @@
 import React, { useEffect, useState } from 'react';
 import { GetServerSidePropsContext, InferGetServerSidePropsType } from 'next';
 import { getVolunteeringList } from '../../src/api';
-import { getAccessLevel } from '../../src/util';
+import { getAccessLevel } from '../../src/util/miscUtilUtil';
 import { darkSwitchGrey } from '../../src/util/cssUtil';
 
 import Grid from '@mui/material/Grid';
