+++ client/src/components/admin/view-feedback.tsx
@@ -1,5 +1,6 @@
 import React, { useEffect, useState } from 'react';
-import { createConnectionErrorPopup, formatDate } from '../../util';
+import { createConnectionErrorPopup } from '../../util/constructors';
+import { formatDate } from '../../util/datetime';
 import { Feedback, PopupEvent } from '../../types';
 import { getFeedback } from '../../api';
