+++ client/src/components/edit/clubs/edit-committee-list.tsx
@@ -1,6 +1,6 @@
 import React, { useEffect, useState } from 'react';
 import type { Control, FieldValues, UseFormRegister, UseFormSetValue } from 'react-hook-form';
-import { createCommittee } from '../../../util';
+import { createCommittee } from '../../../util/constructors';
 import type { Committee } from '../../../types';
 
 import List from '@mui/material/List';
