+++ client/src/components/edit/clubs/edit-exec-list.tsx
@@ -1,7 +1,7 @@
 import React, { useEffect, useState } from 'react';
 import type { Control, FieldValues, UseFormRegister, UseFormSetValue } from 'react-hook-form';
 import type { Exec } from '../../../types';
-import { createExec, createPopupEvent } from '../../../util';
+import { createExec, createPopupEvent } from '../../../util/constructors';
 
 import List from '@mui/material/List';
 import Button from '@mui/material/Button';
