+++ client/src/components/home/home-base.tsx
@@ -1,6 +1,6 @@
 import React, { useEffect, useState } from 'react';
 import { useRouter } from 'next/router';
-import { getParams } from '../../util';
+import { getParams } from '../../util/miscUtilUtil';
 
 import Box from '@mui/material/Box';
 import Drawer from '@mui/material/Drawer';
