+++ client/src/components/home/home-drawer-list.tsx
@@ -7,7 +7,7 @@ import Typography from '@mui/material/Typography';
 import MaterialSymbol from './material-symbol';
 import { getExternalLinks as getExternalLinks } from '../../api';
 import Popup from '../shared/popup';
-import { createPopupEvent } from '../../util';
+import { createPopupEvent } from '../../util/constructors';
 import { darkSwitchGrey } from '../../util/cssUtil';
 import type { ExternalLink } from '../../types';
