+++ client/src/components/shared/delete-button.tsx
@@ -16,7 +16,7 @@ import Popup from './popup';
 import UploadBackdrop from '../edit/shared/upload-backdrop';
 import { PopupEvent, Resource } from '../../types';
 import Cookies from 'universal-cookie';
-import { createPopupEvent } from '../../util';
+import { createPopupEvent } from '../../util/constructors';
 import { deleteClub, deleteEvent, deleteVolunteering } from '../../api';
 
 interface DeleteButtonProps {
